server/requirements.txt
Flask
requests
python-dotenv
flask-cors
chardet
flask-pymongo